refactor: modularização tela lista de progresso
