+++ frontend/lib/features/progress_record/model/Goal.dart
@@ -1,0 +1,15 @@
+class Goal {
+  final String title;
+  final String type;
+  final int? progress;
+  final int? total;
+  final bool isProgressBar;
+
+  Goal({
+    required this.title,
+    required this.type,
+    this.progress,
+    this.total,
+    this.isProgressBar = false,
+  });
+}
