+++ frontend/lib/features/progress_record/model/Progress_record_model.dart
@@ -3,12 +3,16 @@ class ProgressRecordData {
   String? name;
   int? type;
   int? parameter;
+  int? id;
+  String? type_description;
 
   ProgressRecordData({
     this.habitId,
     this.name,
     this.type,
     this.parameter,
+    this.id,
+    this.type_description
   });
 
   ProgressRecordData copyWith({
@@ -16,12 +20,16 @@ class ProgressRecordData {
     int? type,
     String? name,
     int? parameter,
+    int? id,
+    String? type_description
   }) {
     return ProgressRecordData(
       habitId: habitId ?? this.habitId,
       name: name ?? this.name,
       type: type ?? this.type,
-      parameter: parameter ?? this.parameter
+      parameter: parameter ?? this.parameter,
+      id: id ?? this.id,
+      type_description: type_description ?? this.type_description
     );
   }
 
@@ -32,7 +40,9 @@ class ProgressRecordData {
       habitId: $habitId,
       name: $name,
       type: $type,
-      parameter: $parameter
+      parameter: $parameter,
+      id: $id,
+      type_description: $type_description
     )
     ''';
   }
@@ -44,7 +54,20 @@ class ProgressRecordData {
       'habit_id': habitId,
       'name': name,
       'type': type,
-      'parameter': parameter
+      'parameter': parameter,
+      'id': id,
+      'type_description': type_description
     };
+  }
+
+  factory ProgressRecordData.fromJson(Map<String, dynamic> json) {
+    return ProgressRecordData(
+      habitId: json['habit_id'],
+      name: json['name'],
+      type: json['type'],
+      parameter: json['parameter'],
+      id: json['id'],
+      type_description: json['type_description']
+    );
   }
 }
